Add dry-struct and types to conform with basic schema
The idea is for this gem to be used in the datastore (to validate the JSON payload when creating applications), but also it can be used in Apply and Review services, together with the http client, to more easily work with the API responses.

For now this includes some very basic types and structs, to conform to the schema 0.1 that is defined in this same repo.

We have to extend these types, structs and schema until we reach a potential v1.0 we are happy with and includes all the necessary application data from Apply (as we need it back to re-hydrate applications if we want to edit them).