Add case details and codefendants to schema
And to the structs collection.

NOTE: in order to not break current Apply integration, and until we can start sending case details and codefendants in the payload to the datastore, this new JSON object (`case_details`) is not yet `required` in the schema, and is marked as optional in the dry-struct.

Also: some case types values have been renamed. Will update this on Apply side too.