Few minor schema tweaks
Telephone number is not required. It will be made optional on Apply interface soon.

Added a provisional `provider_details` object, required (but for now we just send an empty hash). In preparation for Portal integration on Apply.

Addresses (home and correspondence) are not required. Allow to pass a null object (it will also validate when omitting the attribute completely).