Add address `lookup_id` attribute
And make it explicit the 2 kind of addresses we have (so far), `home_address` and `correspondence_address`.

The `lookup_id` is returned by Ordnance Survey and is needed if we want to re-hydrate the application back to Apply DB and keep the address the user originally selected in the dropdown.