Refactor "people" structs and schema definitions
A bit of reorganisation to better model a top level, basic "Person" containing only the minimum attributes, and then having Codefendant and Applicant (Partner also in the future, maybe even Provider) as individual objects/definitions inheriting from this `Person`.

This is done because Applicant will contain always many more attributes than Codefendant or Partner for instance.

Functionally should not be breaking changes to the schema/structs, the interface is the same.