Update schema to support offence date ranges
After this change is introduced in the datastore, new submitted applications will have to conform to the new format for offence dates (`date_from` and `date_to`).

However to keep compatibility and less disruptive with already submitted applications, the struct counterpart will still support a simple array and will not enforce the content of the objects in the array for the time being.

This will be enforced at a later time.