Enforce new range format, both can't coexist
It seems it is not possible to make both coexist.

New submissions will use new format, but existing submitted applications will have to be migrated to the new format perhaps with an SQL migration.

Another alternative is to wipe up the datastore and start fresh. If we do this, we can enforce some other attributes we made lenient in the past.