Status attribute is no longer required by schema
This makes any submission to the datastore "status-less" and the datastore is now in charge of assigning the status, which by default is `submitted` for any applications submitted by Apply, and in the future Review can update this attribute to `returned` for rejected applications.
This was the current behaviour in reality so nothing changes.

Note: the fixtures are kept with the `status` attribute, as this attribute will continue to be returned from the datastore as it is now. The `struct` definition also enforces this attribute to be present and valid.

Nitpick: we now have a diversion from a "creation schema" and a "read schema", as the creation don't require `status` but the read from datastore requires it.
Currently is not a big pain, it mostly only affects the fixtures, but might grow to something meaningful enough to consider improving.