CRIMAP-331 Add `means_passport` attribute
This serves a similar purpose to the already existing `ioj_passport`, so Review can know on what basis an application has been passported on means.

Also, as Apply now handles this more granularly, we need this information back when reading back applications from the datastore and for rehydration.

Existing applications in datastore lack this attribute, so we can either run a migration task or make the code on Apply/Review resilient with a sensible fallback.

Fixtures and schema updated and fixed some logical inconsistencies.