Tweaks to the schemas, and new definitions
MAAT team has expressed the library they are using can't handle "allOf, anyOf, oneOf".

So as the less disruptive solution, in the MAAT schema we are not going to use these, and instead flatten the "applicant" (which is the only place where these where used). There will be a bit of repetition but is not that bad as that allow us to keep evolving our schema separate from MAAT.

At the same time, I've extracted to shared definitions Interests Of Justice, and Provider details. These are unlikely to change so I think is safe to share these ones between our schema and MAAT schema. We can later change our minds.

Ensured some optional/mandatory attributes are marked as such.