Require `application_type` in schema (#26)
This attribute is always "initial" for MVP, but we prepare the path for post-MVP where other types will be needed.

Fixes the inconsistency between having it declared in the schema (but not required) and not present in any fixture.

Note this is a breaking change for new submitted applications from Apply. This new attribute must be present and with value `initial` (as per enum).

Reading back applications will still work (using `attribute?`) however at some point we want to do an audit of Struct attributes and narrow down which ones should be required and which ones optional. This one should be required (as in present in the JSON stored in the datastore).