Add `passportable` bool attribute to offence object
On the back of this PR on Apply, we now can propagate this from Apply to Datastore, even tho we don't need it just now.

Marking attribute in the Struct as optional, because existing applications in the datastore will not have this.

Not bumping schemas version yet, as there is no need to release a new version just for this.

Will be bumped a bit later once we do some auditing of attributes, structs and schemas that is pending.