Preparation to rationalise and remove some fixtures
We've discussed this a few times, we've let our fixtures grow beyond what is absolutely necessary, and this causes issues with maintaining them, or even knowing if they are used or not.

As a first step I've cleaned up the invalid fixtures and replaced the tests with literally an invalid JSON as it produces the same effect, as what we want to test is the validator fails.

Also, the `LaaCrimeSchemas.fixture` method now can be passed a block. If so, it will yield an already parsed JSON object, which makes it easier to work with some fixtures particularly those with variants, as we don't need fixtures for this, we just need the general fixture, and add/remove/change specific attributes as needed for the tests.

This is not a breaking change, method keeps working as before, but facilitates a path forward to remove some fixtures.