CRIMAP-388 Make the structs more strict
Over the course of the last months we've been evolving the datastore and the submitted JSON documents, with new attributes, or changes to existing ones.

With the objective of not breaking the existing submitted applications every few days, we've been quite lenient in the way the Structs handle the existing data when reading back from the Datastore.

For instance, many attributes were marked as optional, as old data didn't have those. However new submitted applications will always have them now. Etc.

Made changes to the structs to support this. Also a small refactor to the existing structs to have separate `case_details` and `provider_details` and improved testing.

Note these are breaking changes with many of the existing applications in the datastore, so along with this, we would probably want to wipe the DB and start fresh with no baggage.

An additional PR will be raised to the Datastore once the gem new version is tagged.