CRIMAP-433 Refactor return details
These small changes in the schema are needed for a bigger piece of work going on in the datastore to remove the ReturnDetails DB table and use instead a jsonb attribute in the existing CrimeApplications DB table.

The `returned_at` attribute in the schema `return_details` object is redundant as this attribute exists as top-level already, so I've removed it.