CRIMAP-436 MAAT schema changes related to IoJ pass
Removed the `interests_of_justice` array object, because MAAT does not need any of its content, and the only purpose of this attribute would have been to check if the array was empty or not, to trigger IoJ pass logic.

This is now explicitly done via a new boolean attribute, `ioj_bypass`, which will be `true` if the Interests of Justice page was skipped on Apply, meaning we didn't gather IoJ reasons, or `false` otherwise.
It maps more directly to MAAT business rules and domain.