Tidy up structs and fixtures, bump to v1.0.0
Clean up and adapt the structs as we had some temporary workarounds to cope with old datastore data.

We are going to wipe databases so the structs can now be made strict.

Marked as much as possible in the schema as `required` attributes because Apply will always send those ones, just sometimes their value might be null or blank, depending if they are optional or not for the provider to fill.

This will become the v1.0.0 of the gem and the baseline for all applications.